feat: a new feature
Removed the feature Age*Class for Logistic Regression and AdaBoost, and keep the feature for other models. This is due to the impression obtrained from comparing version a and b that Logistic Regression and AdaBoost tend to behave better without Age*Class and Tree models behave better with Age*Class. However, the final result doesn't improve.